use a11y as the default highlighting theme
note that we still use the rstudio theme for the distill website b/c most of the code blocks are yaml and the pandoc yaml tokenization and a11y theme have some interactions that aren't ideal (different coloring of 'distill' in 'distill::distill_article` and different coloring of description from other strings.

I don't think there is anything additional to do here as most of the code is YAJML (so won't seem inconsistent to Distill users).