Add support for references specified within YAML
By default, the final "References" header added if we specify a bibliography in an external .bib file, but not if we manually add references to the "references" field of the header, as described at the start of https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html. This commit aims to fix that.