Correctly detect when run in RStudio subprocess (#456)
Try to detect that render() has been called in RStudio using callr. This could happen using targets for example. This should fix #400

Issue happens because callr is passing the RSTUDIO env var that is set within RStudio so we need another test to rule out that we are in a place where RSTUDIO_VERSION is not set.